requirements.txt
amqp==1.4.9
antigate==1.4.0
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
Django==1.9.1
django-celery==3.1.17
EasyProcess==0.2.1
kombu==3.0.33
Pillow==3.1.1
PyMsgBox==1.0.3
PyScreeze==0.1.8
python3-xlib==0.15
PyTweening==1.0.3
pytz==2015.7
PyVirtualDisplay==0.1.5
redis==2.10.5
requests==2.9.1
selenium==2.49.2
six==1.10.0
wheel==0.24.0
xmltodict==0.9.2
